move sharepoint beta service (#3443)
relocates the sharepoint beta service files into the correct subfolder:
out of m365/api and into connector/sharepoint/api.
