--- src/internal/connector/sharepoint/api/api.go
@@ -1,8 +1,0 @@
-package api
-
-type NameID struct {
-	Name string
-	ID   string
-}
-
-const fetchChannelSize = 5
