+++ src/internal/connector/sharepoint/api/beta_service.go
