+++ src/internal/connector/sharepoint/api/pages_test.go
@@ -10,20 +10,30 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 
+	"github.com/alcionai/corso/src/internal/connector/graph"
 	"github.com/alcionai/corso/src/internal/connector/sharepoint"
 	"github.com/alcionai/corso/src/internal/connector/sharepoint/api"
 	spMock "github.com/alcionai/corso/src/internal/connector/sharepoint/mock"
 	"github.com/alcionai/corso/src/internal/tester"
 	"github.com/alcionai/corso/src/pkg/account"
 	"github.com/alcionai/corso/src/pkg/fault"
-	m365api "github.com/alcionai/corso/src/pkg/services/m365/api"
 )
+
+func createTestBetaService(t *testing.T, credentials account.M365Config) *api.BetaService {
+	adapter, err := graph.CreateAdapter(
+		credentials.AzureTenantID,
+		credentials.AzureClientID,
+		credentials.AzureClientSecret)
+	require.NoError(t, err, clues.ToCore(err))
+
+	return api.NewBetaService(adapter)
+}
 
 type SharePointPageSuite struct {
 	tester.Suite
 	siteID  string
 	creds   account.M365Config
-	service *m365api.BetaService
+	service *api.BetaService
 }
 
 func (suite *SharePointPageSuite) SetupSuite() {
