+++ src/pkg/services/m365/api/drive_pager.go
@@ -13,7 +13,6 @@ import (
 
 	"github.com/alcionai/corso/src/internal/common/ptr"
 	"github.com/alcionai/corso/src/internal/connector/graph"
-	"github.com/alcionai/corso/src/internal/connector/graph/api"
 	onedrive "github.com/alcionai/corso/src/internal/connector/onedrive/consts"
 	"github.com/alcionai/corso/src/pkg/logger"
 )
@@ -66,9 +65,9 @@ func NewItemPager(
 	return res
 }
 
-func (p *driveItemPager) GetPage(ctx context.Context) (api.DeltaPageLinker, error) {
+func (p *driveItemPager) GetPage(ctx context.Context) (DeltaPageLinker, error) {
 	var (
-		resp api.DeltaPageLinker
+		resp DeltaPageLinker
 		err  error
 	)
 
@@ -93,7 +92,7 @@ func (p *driveItemPager) Reset() {
 		Delta()
 }
 
-func (p *driveItemPager) ValuesIn(l api.DeltaPageLinker) ([]models.DriveItemable, error) {
+func (p *driveItemPager) ValuesIn(l DeltaPageLinker) ([]models.DriveItemable, error) {
 	return getValues[models.DriveItemable](l)
 }
 
@@ -135,9 +134,9 @@ type nopUserDrivePageLinker struct {
 
 func (nl nopUserDrivePageLinker) GetOdataNextLink() *string { return nil }
 
-func (p *userDrivePager) GetPage(ctx context.Context) (api.PageLinker, error) {
+func (p *userDrivePager) GetPage(ctx context.Context) (PageLinker, error) {
 	var (
-		resp api.PageLinker
+		resp PageLinker
 		err  error
 	)
 
@@ -163,7 +162,7 @@ func (p *userDrivePager) SetNext(link string) {
 	p.builder = users.NewItemDrivesRequestBuilder(link, p.gs.Adapter())
 }
 
-func (p *userDrivePager) ValuesIn(l api.PageLinker) ([]models.Driveable, error) {
+func (p *userDrivePager) ValuesIn(l PageLinker) ([]models.Driveable, error) {
 	nl, ok := l.(*nopUserDrivePageLinker)
 	if !ok || nl == nil {
 		return nil, clues.New(fmt.Sprintf("improper page linker struct for user drives: %T", l))
@@ -212,9 +211,9 @@ func NewSiteDrivePager(
 	return res
 }
 
-func (p *siteDrivePager) GetPage(ctx context.Context) (api.PageLinker, error) {
+func (p *siteDrivePager) GetPage(ctx context.Context) (PageLinker, error) {
 	var (
-		resp api.PageLinker
+		resp PageLinker
 		err  error
 	)
 
@@ -230,7 +229,7 @@ func (p *siteDrivePager) SetNext(link string) {
 	p.builder = sites.NewItemDrivesRequestBuilder(link, p.gs.Adapter())
 }
 
-func (p *siteDrivePager) ValuesIn(l api.PageLinker) ([]models.Driveable, error) {
+func (p *siteDrivePager) ValuesIn(l PageLinker) ([]models.Driveable, error) {
 	return getValues[models.Driveable](l)
 }
 
@@ -240,9 +239,9 @@ func (p *siteDrivePager) ValuesIn(l api.PageLinker) ([]models.Driveable, error) 
 
 // DrivePager pages through different types of drive owners
 type DrivePager interface {
-	GetPage(context.Context) (api.PageLinker, error)
+	GetPage(context.Context) (PageLinker, error)
 	SetNext(nextLink string)
-	ValuesIn(api.PageLinker) ([]models.Driveable, error)
+	ValuesIn(PageLinker) ([]models.Driveable, error)
 }
 
 // GetAllDrives fetches all drives for the given pager
@@ -262,7 +261,7 @@ func GetAllDrives(
 	for {
 		var (
 			err  error
-			page api.PageLinker
+			page PageLinker
 		)
 
 		// Retry Loop for Drive retrieval. Request can timeout
@@ -311,7 +310,7 @@ func GetAllDrives(
 // Helpers
 // ---------------------------------------------------------------------------
 
-func getValues[T any](l api.PageLinker) ([]T, error) {
+func getValues[T any](l PageLinker) ([]T, error) {
 	page, ok := l.(interface{ GetValue() []T })
 	if !ok {
 		return nil, clues.New("page does not comply with GetValue() interface").With("page_item_type", fmt.Sprintf("%T", l))
